HSEARCH-3729 Move MassIndexingMonitor to the same package as MassIndexer
This feature is not complex enough to warrant a dedicated package.
