+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/massindexing/MassIndexer.java
@@ -9,7 +9,6 @@ package org.hibernate.search.mapper.orm.massindexing;
 import java.util.concurrent.CompletableFuture;
 
 import org.hibernate.CacheMode;
-import org.hibernate.search.mapper.orm.massindexing.monitor.MassIndexingMonitor;
 
 /**
  * A MassIndexer is useful to rebuild the indexes from the
